AMB-1735 refactor pds_service
